functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.4.0
facebook-business>=20.0.0
google-generativeai>=0.8.0
python-telegram-bot>=21.0
sendgrid>=6.11.0
twilio>=9.0.0
cryptography>=42.0.0
requests>=2.31.0
numpy>=1.26.0
reportlab>=4.1.0
matplotlib>=3.8.0
